InputHandlerの修正
