+++ MG-Force-Project/Assets/Scenes/TestScene/MoveTest.unity
@@ -191,6 +191,39 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: be7acc96bae851a439c97c45c3244e20, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &143286818
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 143286807}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9319f62d217bd5543ac84a7b31057ff5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!95 &143286822
+Animator:
+  serializedVersion: 5
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 143286807}
+  m_Enabled: 1
+  m_Avatar: {fileID: 0}
+  m_Controller: {fileID: 0}
+  m_CullingMode: 0
+  m_UpdateMode: 0
+  m_ApplyRootMotion: 0
+  m_LinearVelocityBlending: 0
+  m_StabilizeFeet: 0
+  m_WarningMessage: 
+  m_HasTransformHierarchy: 1
+  m_AllowConstantClipSamplingOptimization: 1
+  m_KeepAnimatorStateOnDisable: 0
+  m_WriteDefaultValuesOnDisable: 0
 --- !u!1001 &209531782
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -270,6 +303,14 @@ PrefabInstance:
         type: 3}
       insertIndex: -1
       addedObject: {fileID: 143286817}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 697447828b80a894c9860d0a52d90fc5,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 143286818}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 697447828b80a894c9860d0a52d90fc5,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 143286822}
   m_SourcePrefab: {fileID: 100100000, guid: 697447828b80a894c9860d0a52d90fc5, type: 3}
 --- !u!1 &330585543
 GameObject:
