--- MG-Force-Project/Assets/Scripts/System/InputSystem/PlayerInputHandler1.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 13dda5b242be113418eab9e15a0ce980
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
